feat: 가족 탭 API 연동 (#79)
* feat : 가족 탭 api연동

* fix: 가족 탭 API 연동 안정화 및 버그 수정

* feat: 카카오톡 sdk추가

* fix: 코드래빗 피드백 반영 및 버그 수정

+++ app/src/main/java/com/example/senior_on/data/remote/dto/FamilyDtos.kt
@@ -15,12 +15,12 @@ data class FamilyPhotoItemResponse(
 )
 data class FamilyPhotoCursorResponse(val createdAt: String?, val familyPhotoId: Long?)
 data class FamilyPhotoListResponse(
-    val photos: List<FamilyPhotoItemResponse>?, val totalCount: Int?,
+    val photos: List<FamilyPhotoItemResponse>?, val totalCount: Long?,
     val nextCursor: FamilyPhotoCursorResponse?, val hasNext: Boolean?
 )
 data class FamilyPhotoAlbumResponse(
     val uploaderUserId: Long?, val uploaderName: String?, val latestPhotoUrl: String?,
-    val photoCount: Int?, val hasNewPhotos: Boolean?
+    val photoCount: Long?, val hasNewPhotos: Boolean?
 )
 data class FamilyMemberResponse(
     val usersId: Long?, val name: String?, val role: String?,
@@ -32,4 +32,4 @@ data class FamilyHomeResponse(
     val recentUploaderProfileImageUrls: List<String>?,
     val recentPhotos: List<FamilyPhotoItemResponse>?
 )
-data class FamilyCodeResponse(val familyCode: String?, val familyMemberCount: Int?)
+data class FamilyCodeResponse(val familyCode: String?, val familyMemberCount: Long?)
